+++ src/app/cart/components/index.ts
@@ -1,0 +1,2 @@
+export { CartListComponent } from './cart-list/cart-list.component';
+export { CartItemComponent } from './cart-item/cart-item.component';