+++ src/app/cart/services/cart.service.ts
@@ -1,30 +1,32 @@
 import { Injectable } from '@angular/core';
 import { Subject } from 'rxjs';
-import { Product } from '../../products/models/product.model';
+import { CartProduct, Product } from '../../shared/models/product.model';
 
 @Injectable({
   providedIn: 'root'
 })
 export class CartService {
-  addedProducts: Subject<Product[]> = new Subject<Product[]>();
+  addedProducts: Subject<CartProduct[]> = new Subject<CartProduct[]>();
 
-  private products: Product[] = [];
+  private products: CartProduct[] = [];
 
   constructor() { }
 
   addProductToList(product: Product): void {
     const productInCart = this.products.find((item) => {
-      return item.id === product.id;
+      return item.product.id === product.id;
     });
-    if (productInCart?.quantity) {
+    if (productInCart) {
       productInCart.quantity = productInCart.quantity + 1;
+      productInCart.totalPrice = productInCart.product.price * productInCart.quantity;
       this.products = this.products.map((item) => {
-        return productInCart.id === item.id ? productInCart : item;
+        return productInCart.product.id === item.product.id ? productInCart : item;
       });
     } else {
       this.products = [...this.products, {
-        ...product,
+        product,
         quantity: 1,
+        totalPrice: product.price
       }];
     }
     this.addedProducts.next(this.products);
@@ -32,20 +34,23 @@ export class CartService {
 
   removeProduct(product: Product, removeAll: boolean = false): void {
     const productInCart = this.products.find((item) => {
-      return item.id === product.id;
+      return item.product.id === product.id;
     });
     if (productInCart?.quantity && productInCart.quantity > 1 && !removeAll) {
+      const newQuantity = productInCart.quantity - 1;
+      const newRecord = {
+        ...productInCart,
+        quantity: newQuantity,
+        totalPrice: productInCart.product.price * newQuantity
+      };
       this.products = this.products.map((item) => {
-        return productInCart.id === item.id
-        ? {
-          ...productInCart,
-          quantity: productInCart.quantity ? productInCart.quantity - 1 : 1,
-        }
+        return productInCart.product.id === item.product.id
+        ? newRecord
         : item;
       });
     } else {
       this.products = this.products.filter((item) => {
-        return item.id !== product.id;
+        return item.product.id !== product.id;
       });
     }
     this.addedProducts.next(this.products);