+++ src/app/products/components/index.ts
@@ -1,0 +1,2 @@
+export { ProductListComponent } from './product-list/product-list.component';
+export { ProductComponent } from './product/product.component';