+++ src/app/products/products.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { ProductComponent, ProductListComponent } from './components';
+import { SharedModule } from '../shared/shared.module';
+
+@NgModule({
+  declarations: [ProductComponent, ProductListComponent],
+  imports: [
+    CommonModule,
+    SharedModule,
+  ],
+  exports: [ProductComponent, ProductListComponent],
+})
+export class ProductsModule { }