+++ src/app/products/services/products.service.ts
@@ -1,14 +1,14 @@
 import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
 import { Categories } from '../enums/categories.enum';
-import { Product } from '../models/product.model';
+import { Product } from '../../shared/models/product.model';
 
 const products: Product[] = [
   {
     id: 1,
     name: 'HP laptop',
     description: 'Some cool laptop',
-    price: 1111,
+    price: 1749,
     category: 0,
     isAvailable: true,
   },
@@ -16,7 +16,7 @@ const products: Product[] = [
     id: 2,
     name: 'HP manual',
     description: 'Some book',
-    price: 22,
+    price: 19,
     category: 2,
     isAvailable: true,
   },
@@ -24,7 +24,7 @@ const products: Product[] = [
     id: 3,
     name: 'Ball',
     description: 'Some ball',
-    price: 33,
+    price: 31,
     category: 3,
     isAvailable: false,
   }