+++ src/app/shared/directives/highlight.directive.ts
@@ -1,0 +1,21 @@
+import { Directive, HostBinding, HostListener } from '@angular/core';
+
+@Directive({
+  selector: '[appHighlight]'
+})
+export class HighlightDirective {
+
+  @HostBinding('class')
+  attrClass = 'highlight';
+
+  @HostListener('mouseenter', ['$event'])
+  enter(event: Event): void {
+    this.attrClass = 'highlight highlight-active';
+  }
+
+  @HostListener('mouseleave', ['$event'])
+  leave(event: Event): void {
+    this.attrClass = 'highlight';
+  }
+
+}