+++ src/app/shared/models/product.model.ts
@@ -1,0 +1,16 @@
+import { Categories } from '../../products/enums/categories.enum';
+
+export interface Product {
+  id: number;
+  name: string;
+  description: string;
+  price: number;
+  category: Categories;
+  isAvailable: boolean;
+}
+
+export interface CartProduct {
+  product: Product;
+  quantity: number;
+  totalPrice: number;
+}